requirements.txt
fastapi
uvicorn
requests
pydantic
psycopg2-binary
python-dotenv
